# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.MobileOrderDetailResponse:
    properties:
      address:
        type: string
      buyer:
        type: string
      cancelled_at:
        type: string
      cancelled_by:
        type: string
      changed_at:
        type: string
      changed_by:
        type: string
      channel:
        type: string
      courier:
        type: string
      created_at:
        type: string
      event_status:
        type: string
      id:
        type: integer
      order_details:
        items:
          $ref: '#/definitions/controllers.MobileOrderDetailWithProduct'
        type: array
      order_ginee_id:
        type: string
      pending_at:
        type: string
      pending_by:
        type: string
      picked_at:
        type: string
      picked_by:
        type: string
      processing_status:
        type: string
      sent_before:
        type: string
      store:
        type: string
      tracking:
        type: string
      updated_at:
        type: string
    type: object
  controllers.MobileOrderDetailWithProduct:
    properties:
      barcode:
        type: string
      id:
        type: integer
      image:
        type: string
      location:
        type: string
      product:
        allOf:
        - $ref: '#/definitions/models.ProductResponse'
        description: Related data
      product_name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      variant:
        type: string
    type: object
  controllers.MobileOrderListResponse:
    properties:
      address:
        type: string
      buyer:
        type: string
      cancelled_at:
        type: string
      cancelled_by:
        type: string
      changed_at:
        type: string
      changed_by:
        type: string
      channel:
        type: string
      courier:
        type: string
      created_at:
        type: string
      event_status:
        type: string
      id:
        type: integer
      order_details:
        items:
          $ref: '#/definitions/controllers.MobileOrderDetailWithProduct'
        type: array
      order_ginee_id:
        type: string
      pending_at:
        type: string
      pending_by:
        type: string
      picked_at:
        type: string
      picked_by:
        type: string
      processing_status:
        type: string
      sent_before:
        type: string
      store:
        type: string
      tracking:
        type: string
      updated_at:
        type: string
    type: object
  controllers.MobileOrdersListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/controllers.MobileOrderListResponse'
        type: array
      pagination:
        $ref: '#/definitions/utilities.PaginationResponse'
    type: object
...
  models.OrderDetailResponse:
    properties:
      id:
        type: integer
      product:
        allOf:
        - $ref: '#/definitions/models.ProductResponse'
        description: Related data
      product_name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      variant:
        type: string
    type: object
  models.OrderResponse:
    properties:
      address:
        type: string
      buyer:
        type: string
      cancelled_at:
        type: string
      cancelled_by:
        type: string
      changed_at:
        type: string
      changed_by:
        type: string
      channel:
        type: string
      complained:
        type: boolean
      courier:
        type: string
      created_at:
        type: string
      event_status:
        type: string
      id:
        type: integer
      order_details:
        description: Related data
        items:
          $ref: '#/definitions/models.OrderDetailResponse'
        type: array
      order_ginee_id:
        type: string
      pending_at:
        type: string
      pending_by:
        type: string
      picked_at:
        type: string
      picked_by:
        type: string
      processing_status:
        type: string
      sent_before:
        type: string
      store:
        type: string
      tracking:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ProductResponse:
    properties:
      barcode:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      location:
        type: string
      name:
        type: string
      sku:
        type: string
      updated_at:
        type: string
      variant:
        type: string
    type: object
...
  utilities.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  utilities.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@livotech.com
    name: API Support
  description: A comprehensive user management backend service with JWT authentication
    and role-based access control
  title: Livotech Backend Service API
  version: "1.0"
paths:
...
  /api/mobile/orders:
    get:
      consumes:
      - application/json
      description: Get list of all orders with "ready to pick" processing status,
        Optional search by order ID or tracking number.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by order Ginee ID or tracking number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.MobileOrdersListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Get all orders by mobile
      tags:
      - mobile-orders
  /api/mobile/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get specific order details with product location and barcode joined
        by SKU.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.MobileOrderDetailResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Get order by ID by mobile
      tags:
      - mobile-orders
  /api/mobile/orders/{id}/complete:
    put:
      consumes:
      - application/json
      description: Change order processing status from "picking process" to "picking
        complete" and create pick order records
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utilities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Complete picking process by mobile
      tags:
      - mobile-orders
  /api/mobile/orders/{id}/pending:
    put:
      consumes:
      - application/json
      description: Change order processing status from "picking process" to "pending
        picking" and unassign picker
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utilities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Pending picking process by mobile
      tags:
      - mobile-orders
  /api/mobile/orders/{id}/pick:
    put:
      consumes:
      - application/json
      description: Change order processing status from "ready to pick" to "picking
        process" and assign to current picker
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utilities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Pick an order for processing by mobile
      tags:
      - mobile-orders
  /api/mobile/orders/my-picking:
    get:
      consumes:
      - application/json
      description: 'Get list of orders currently being picked by the logged-in user
        (processing status: "picking process")'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Get my ongoing picking orders by mobile
      tags:
      - mobile-orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"